+++ src/main/java/com/ceos/bankids/service/FamilyUserService.java
@@ -3,7 +3,6 @@ package com.ceos.bankids.service;
 import com.ceos.bankids.domain.Family;
 import com.ceos.bankids.domain.FamilyUser;
 import com.ceos.bankids.domain.User;
-import com.ceos.bankids.dto.KidListDTO;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
@@ -13,9 +12,9 @@ public interface FamilyUserService {
 
     public void checkIfFamilyExists(User user);
 
-    public void postNewFamilyUser(Family family, User user);
+    public void createFamilyUser(Family family, User user);
 
-    public void leavePreviousFamily(User user);
+    public void leavePreviousFamilyIfExists(User user);
 
     public Optional<FamilyUser> findByUserNullable(User user);
 
@@ -23,13 +22,11 @@ public interface FamilyUserService {
 
     public FamilyUser findByUserAndCheckCode(User user, String code);
 
-    public List<FamilyUser> checkFamilyUserList(Family family, User user);
+    public List<FamilyUser> getFamilyUserList(Family family, User user);
 
     public void deleteFamilyUser(FamilyUser familyUser);
 
     public List<FamilyUser> getFamilyUserListExclude(Family family, User user);
-
-    public List<KidListDTO> getKidListFromFamily(FamilyUser familyUser);
 
     public User getContractUser(User user, Boolean isMom);
